Move deleteRootComment into the right spot
